refactor: move pypi requirement to crate (#3689)
