+++ crates/pixi_manifest/Cargo.toml
@@ -19,8 +19,9 @@ pathdiff = { workspace = true }
 pep440_rs = { workspace = true }
 pep508_rs = { workspace = true }
 pixi_consts = { workspace = true }
-pixi_git = { workspace = true }
+pixi_pypi_spec = { workspace = true }
 pixi_spec = { workspace = true }
+pixi_spec_containers = { workspace = true }
 pixi_toml = { workspace = true }
 regex = { workspace = true }
 serde = { workspace = true }
@@ -57,6 +58,7 @@ fancy_display = { workspace = true }
 glob = "0.3.2"
 insta = { workspace = true, features = ["yaml"] }
 miette = { workspace = true, features = ["fancy-no-backtrace"] }
+pixi_test_utils = { workspace = true }
 rstest = { workspace = true }
 serde_json = { workspace = true, features = ["preserve_order"] }
 tempfile = { workspace = true }
