+++ crates/pixi_manifest/src/lib.rs
@@ -30,7 +30,7 @@ pub use activation::Activation;
 pub use build_system::BuildBackend;
 pub use build_system::PackageBuild;
 pub use channel::PrioritizedChannel;
-pub use dependencies::{CondaDependencies, Dependencies, PyPiDependencies};
+pub use dependencies::{CondaDependencies, PyPiDependencies};
 pub use discovery::{
     DiscoveryStart, ExplicitManifestError, LoadManifestsError, Manifests, WorkspaceDiscoverer,
     WorkspaceDiscoveryError,
@@ -49,7 +49,6 @@ pub use manifests::{
 use miette::Diagnostic;
 pub use package::Package;
 pub use preview::{KnownPreviewFeature, Preview};
-pub use pypi::pypi_requirement::PyPiRequirement;
 use rattler_conda_types::Platform;
 pub use s3::S3Options;
 pub use spec_type::SpecType;
