+++ crates/pixi_manifest/src/manifests/document.rs
@@ -1,18 +1,17 @@
 use std::{fmt, str::FromStr, sync::Arc};
 
+use crate::{
+    FeatureName, LibCSystemRequirement, ManifestKind, ManifestProvenance, PypiDependencyLocation,
+    SpecType, SystemRequirements, Task, TomlError, manifests::table_name::TableName,
+    toml::TomlDocument, utils::WithSourceCode,
+};
 use miette::{Diagnostic, NamedSource};
 use pixi_consts::consts;
+use pixi_pypi_spec::{PixiPypiSpec, PypiPackageName};
 use pixi_spec::PixiSpec;
 use rattler_conda_types::{PackageName, Platform};
 use thiserror::Error;
 use toml_edit::{Array, DocumentMut, Item, Table, Value, value};
-
-use crate::{
-    FeatureName, LibCSystemRequirement, ManifestKind, ManifestProvenance, PyPiRequirement,
-    PypiDependencyLocation, SpecType, SystemRequirements, Task, TomlError,
-    manifests::table_name::TableName, pypi::PyPiPackageName, toml::TomlDocument,
-    utils::WithSourceCode,
-};
 
 /// Discriminates between a 'pixi.toml' and a 'pyproject.toml' manifest.
 #[derive(Debug, Clone)]
@@ -231,7 +230,7 @@ impl ManifestDocument {
     /// If will be a no-op if the dependency is not found.
     pub fn remove_pypi_dependency(
         &mut self,
-        dep: &PyPiPackageName,
+        dep: &PypiPackageName,
         platform: Option<Platform>,
         feature_name: &FeatureName,
     ) -> Result<(), TomlError> {
@@ -249,7 +248,7 @@ impl ManifestDocument {
                             .unwrap_or("")
                             .parse()
                             .expect("should be a valid pep508 dependency");
-                        let name = PyPiPackageName::from_normalized(req.name);
+                        let name = PypiPackageName::from_normalized(req.name);
                         name != *dep
                     });
                     if array.is_empty() {
@@ -352,7 +351,7 @@ impl ManifestDocument {
         // before adding it back
         if matches!(self, ManifestDocument::PyProjectToml(_)) {
             self.remove_pypi_dependency(
-                &PyPiPackageName::from_normalized(requirement.name.clone()),
+                &PypiPackageName::from_normalized(requirement.name.clone()),
                 platform,
                 feature_name,
             )?;
@@ -370,7 +369,7 @@ impl ManifestDocument {
             || editable.is_some_and(|e| e)
         {
             let mut pypi_requirement =
-                PyPiRequirement::try_from(requirement.clone()).map_err(Box::new)?;
+                PixiPypiSpec::try_from(requirement.clone()).map_err(Box::new)?;
             if let Some(editable) = editable {
                 pypi_requirement.set_editable(editable);
             }
@@ -440,7 +439,7 @@ impl ManifestDocument {
     /// found, or `None` if it is not found in any of the checked sections.
     pub fn pypi_dependency_location(
         &self,
-        package_name: &PyPiPackageName,
+        package_name: &PypiPackageName,
         platform: Option<Platform>,
         feature_name: &FeatureName,
     ) -> Option<PypiDependencyLocation> {
