+++ crates/pixi_manifest/src/pypi/mod.rs
@@ -1,4 +1,1 @@
 pub mod pypi_options;
-pub mod pypi_requirement;
-pub mod pypi_requirement_types;
-pub use pypi_requirement_types::{PyPiPackageName, VersionOrStar};
