+++ crates/pixi_manifest/src/toml/environment.rs
@@ -59,12 +59,12 @@ impl<'de> toml_span::Deserialize<'de> for TomlEnvironmentList {
 
 #[cfg(test)]
 mod test {
+    use super::*;
+    use crate::toml::FromTomlStr;
     use assert_matches::assert_matches;
     use insta::assert_snapshot;
+    use pixi_test_utils::format_parse_error;
     use toml_span::{DeserError, Value, de_helpers::TableHelper};
-
-    use super::*;
-    use crate::{toml::FromTomlStr, utils::test_utils::format_parse_error};
 
     #[derive(Debug)]
     struct TopLevel {
