+++ crates/pixi_manifest/src/toml/feature.rs
@@ -6,9 +6,9 @@ use rattler_conda_types::Platform;
 use toml_span::{DeserError, Spanned, Value, de_helpers::TableHelper};
 
 use crate::{
-    Activation, Feature, FeatureName, PyPiRequirement, SystemRequirements, TargetSelector, Targets,
-    Task, TaskName, TomlError, Warning, WithWarnings,
-    pypi::{PyPiPackageName, pypi_options::PypiOptions},
+    Activation, Feature, FeatureName, SystemRequirements, TargetSelector, Targets, Task, TaskName,
+    TomlError, Warning, WithWarnings,
+    pypi::pypi_options::PypiOptions,
     toml::{
         PlatformSpan, TomlPrioritizedChannel, TomlTarget, TomlWorkspace,
         create_unsupported_selector_warning, platform::TomlPlatform, preview::TomlPreview,
@@ -17,6 +17,7 @@ use crate::{
     utils::{PixiSpanned, package_map::UniquePackageMap},
     workspace::ChannelPriority,
 };
+use pixi_pypi_spec::{PixiPypiSpec, PypiPackageName};
 
 #[derive(Debug)]
 pub struct TomlFeature {
@@ -28,7 +29,7 @@ pub struct TomlFeature {
     pub dependencies: Option<PixiSpanned<UniquePackageMap>>,
     pub host_dependencies: Option<PixiSpanned<UniquePackageMap>>,
     pub build_dependencies: Option<PixiSpanned<UniquePackageMap>>,
-    pub pypi_dependencies: Option<IndexMap<PyPiPackageName, PyPiRequirement>>,
+    pub pypi_dependencies: Option<IndexMap<PypiPackageName, PixiPypiSpec>>,
 
     /// Additional information to activate an environment.
     pub activation: Option<Activation>,
