+++ crates/pixi_manifest/src/toml/manifest.rs
@@ -4,8 +4,24 @@ use std::{
     path::{Path, PathBuf},
 };
 
+use crate::{
+    Activation, Environment, EnvironmentName, Environments, Feature, FeatureName,
+    KnownPreviewFeature, SolveGroups, SystemRequirements, TargetSelector, Targets, Task, TaskName,
+    TomlError, Warning, WithWarnings, WorkspaceManifest,
+    environment::EnvironmentIdx,
+    error::{FeatureNotEnabled, GenericError},
+    manifests::PackageManifest,
+    pypi::pypi_options::PypiOptions,
+    toml::{
+        ExternalPackageProperties, PlatformSpan, TomlFeature, TomlPackage, TomlTarget,
+        TomlWorkspace, create_unsupported_selector_warning, environment::TomlEnvironmentList,
+        task::TomlTask,
+    },
+    utils::{PixiSpanned, package_map::UniquePackageMap},
+};
 use indexmap::IndexMap;
 use miette::LabeledSpan;
+use pixi_pypi_spec::{PixiPypiSpec, PypiPackageName};
 use pixi_toml::{Same, TomlHashMap, TomlIndexMap, TomlWith};
 use rattler_conda_types::{Platform, Version};
 use toml_span::{
@@ -14,22 +30,6 @@ use toml_span::{
     value::ValueInner,
 };
 use url::Url;
-
-use crate::{
-    Activation, Environment, EnvironmentName, Environments, Feature, FeatureName,
-    KnownPreviewFeature, PyPiRequirement, SolveGroups, SystemRequirements, TargetSelector, Targets,
-    Task, TaskName, TomlError, Warning, WithWarnings, WorkspaceManifest,
-    environment::EnvironmentIdx,
-    error::{FeatureNotEnabled, GenericError},
-    manifests::PackageManifest,
-    pypi::{PyPiPackageName, pypi_options::PypiOptions},
-    toml::{
-        ExternalPackageProperties, PlatformSpan, TomlFeature, TomlPackage, TomlTarget,
-        TomlWorkspace, create_unsupported_selector_warning, environment::TomlEnvironmentList,
-        task::TomlTask,
-    },
-    utils::{PixiSpanned, package_map::UniquePackageMap},
-};
 
 /// Raw representation of a pixi manifest. This is the deserialized form of the
 /// manifest without any validation logic applied.
@@ -43,7 +43,7 @@ pub struct TomlManifest {
     pub dependencies: Option<PixiSpanned<UniquePackageMap>>,
     pub host_dependencies: Option<PixiSpanned<UniquePackageMap>>,
     pub build_dependencies: Option<PixiSpanned<UniquePackageMap>>,
-    pub pypi_dependencies: Option<PixiSpanned<IndexMap<PyPiPackageName, PyPiRequirement>>>,
+    pub pypi_dependencies: Option<PixiSpanned<IndexMap<PypiPackageName, PixiPypiSpec>>>,
 
     /// Additional information to activate an environment.
     pub activation: Option<PixiSpanned<Activation>>,
@@ -575,13 +575,10 @@ pub struct ExternalWorkspaceProperties {
 
 #[cfg(test)]
 mod test {
-    use insta::assert_snapshot;
-
     use super::*;
-    use crate::{
-        toml::FromTomlStr,
-        utils::test_utils::{expect_parse_warnings, format_parse_error},
-    };
+    use crate::{toml::FromTomlStr, utils::test_utils::expect_parse_warnings};
+    use insta::assert_snapshot;
+    use pixi_test_utils::format_parse_error;
 
     /// A helper function that generates a snapshot of the error message when
     /// parsing a manifest TOML. The error is returned.
