+++ crates/pixi_manifest/src/toml/pypi_options.rs
@@ -193,10 +193,10 @@ impl<'de> toml_span::Deserialize<'de> for NoBuildIsolation {
 
 #[cfg(test)]
 mod test {
-    use insta::{assert_debug_snapshot, assert_snapshot};
-
     use super::*;
-    use crate::{toml::FromTomlStr, utils::test_utils::format_parse_error};
+    use crate::toml::FromTomlStr;
+    use insta::{assert_debug_snapshot, assert_snapshot};
+    use pixi_test_utils::format_parse_error;
 
     #[test]
     fn test_empty() {
