+++ crates/pixi_pypi_spec/src/version_or_star.rs
@@ -1,55 +1,9 @@
-use std::{
-    borrow::Borrow,
-    fmt::{Display, Formatter, Write},
-    str::FromStr,
-};
-
 use pep440_rs::VersionSpecifiers;
-use pep508_rs::{InvalidNameError, PackageName};
 use pixi_toml::TomlFromStr;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
+use std::fmt::{Display, Formatter, Write};
+use std::str::FromStr;
 use toml_span::{DeserError, Value};
-
-#[derive(Debug, Clone, Eq, PartialEq, Hash)]
-/// A package name for PyPI that also stores the source version of the name.
-pub struct PyPiPackageName {
-    source: String,
-    normalized: PackageName,
-}
-
-impl Borrow<PackageName> for PyPiPackageName {
-    fn borrow(&self) -> &PackageName {
-        &self.normalized
-    }
-}
-
-impl FromStr for PyPiPackageName {
-    type Err = InvalidNameError;
-
-    fn from_str(name: &str) -> Result<Self, Self::Err> {
-        Ok(Self {
-            source: name.to_string(),
-            normalized: PackageName::from_str(name)?,
-        })
-    }
-}
-
-impl PyPiPackageName {
-    pub fn from_normalized(normalized: PackageName) -> Self {
-        Self {
-            source: normalized.to_string(),
-            normalized,
-        }
-    }
-
-    pub fn as_normalized(&self) -> &PackageName {
-        &self.normalized
-    }
-
-    pub fn as_source(&self) -> &str {
-        &self.source
-    }
-}
 
 /// The pep crate does not support "*" as a version specifier, so we need to
 /// handle it ourselves.
