+++ crates/pixi_test_utils/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+license.workspace = true
+name = "pixi_test_utils"
+readme.workspace = true
+repository.workspace = true
+version = "0.1.0"
+
+[dependencies]
+itertools = { workspace = true }
+miette = { workspace = true, features = ["fancy-no-backtrace"] }
