+++ crates/pixi_toml/Cargo.toml
@@ -14,6 +14,8 @@ digest = { workspace = true }
 hex = { workspace = true }
 indexmap = { workspace = true }
 itertools = { workspace = true }
+miette = { workspace = true }
+strsim = { workspace = true }
 strum = { workspace = true }
 toml-span = { workspace = true }
