+++ crates/pixi_uv_conversions/Cargo.toml
@@ -19,6 +19,7 @@ pep440_rs = { workspace = true }
 pep508_rs = { workspace = true }
 pixi_git = { workspace = true }
 pixi_manifest = { workspace = true }
+pixi_pypi_spec = { workspace = true }
 pixi_record = { workspace = true }
 pixi_spec = { workspace = true }
 rattler_lock = { workspace = true }
