+++ src/cli/has_specs.rs
@@ -1,10 +1,9 @@
+use crate::Workspace;
 use indexmap::IndexMap;
 use miette::IntoDiagnostic;
 use pep508_rs::Requirement;
+use pixi_pypi_spec::PypiPackageName;
 use rattler_conda_types::{MatchSpec, PackageName, ParseStrictness};
-
-use crate::Workspace;
-use pixi_manifest::pypi::PyPiPackageName;
 
 /// A trait to facilitate extraction of packages data from arguments
 pub(crate) trait HasSpecs {
@@ -28,12 +27,12 @@ pub(crate) trait HasSpecs {
     fn pypi_deps(
         &self,
         project: &Workspace,
-    ) -> miette::Result<IndexMap<PyPiPackageName, Requirement>> {
+    ) -> miette::Result<IndexMap<PypiPackageName, Requirement>> {
         self.packages()
             .iter()
             .map(|package| {
                 let dep = Requirement::parse(package, project.root()).into_diagnostic()?;
-                let name = PyPiPackageName::from_normalized(dep.clone().name);
+                let name = PypiPackageName::from_normalized(dep.clone().name);
                 Ok((name, dep))
             })
             .collect()
